+++ src/main/resources/db/migration/V10__create-table-tipo_usuario_permissao.sql
@@ -1,8 +1,8 @@
 -- Criação da tabela que liga usuários e permissões
 CREATE TABLE tipo_usuario_permissao
 (
-    id_tipo_usuario_permissao INT AUTO_INCREMENT PRIMARY KEY,
-    id_tipo_usuario           INT          NOT NULL,
+    id_tipo_usuario_permissao BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id_tipo_usuario           BIGINT       NOT NULL,
     end_point                 VARCHAR(255) NOT NULL,
-    is_active                 TINYINT(1) DEFAULT 1
+    is_active                 TINYINT DEFAULT 1
 );
