+++ src/main/resources/db/migration/V12__create-table-endereco.sql
@@ -1,14 +1,14 @@
 -- Criação da tabela de endereço
 CREATE TABLE endereco
 (
-    id_endereco INT AUTO_INCREMENT PRIMARY KEY,
-    id_paciente INT          NOT NULL,
+    id_endereco BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id_paciente BIGINT       NOT NULL,
     logradouro  VARCHAR(255) NOT NULL,
     numero      VARCHAR(10),
     complemento VARCHAR(255),
     bairro      VARCHAR(100),
     cidade      VARCHAR(100) NOT NULL,
-    id_uf       INT          NOT NULL,
+    id_uf       BIGINT       NOT NULL,
     cep         VARCHAR(8)   NOT NULL,
     referencia  VARCHAR(255)
 )
