+++ src/main/resources/db/migration/V14__create-table-arquivo.sql
@@ -1,16 +1,16 @@
 -- Criação da tabela de arquivos
 CREATE TABLE arquivo
 (
-    id_arquivo      INT AUTO_INCREMENT PRIMARY KEY,
+    id_arquivo      BIGINT AUTO_INCREMENT PRIMARY KEY,
     entidade        VARCHAR(255) NOT NULL,
-    id_entidade     INT          NOT NULL,
-    id_tipo_arquivo INT,
+    id_entidade     BIGINT       NOT NULL,
+    id_tipo_arquivo BIGINT,
     nome_arquivo    VARCHAR(255) NOT NULL,
     caminho_arquivo VARCHAR(255) NOT NULL,
     tamanho         BIGINT,
     formato         VARCHAR(50),
     data_upload     DATETIME DEFAULT CURRENT_TIMESTAMP,
     metadados       JSON,
-    is_active       TINYINT(1) DEFAULT 1,
+    is_active       TINYINT  DEFAULT 1,
     INDEX (entidade, id_entidade)
 );
