+++ src/main/resources/db/migration/V16__create-table-plano_saude.sql
@@ -1,8 +1,8 @@
 -- Criação da tabela de Planos de Saúde
 CREATE TABLE plano_saude
 (
-    id_plano_saude INT AUTO_INCREMENT PRIMARY KEY,
-    descricao           VARCHAR(255) NOT NULL,
+    id_plano_saude BIGINT AUTO_INCREMENT PRIMARY KEY,
+    descricao      VARCHAR(255) NOT NULL,
     is_active      TINYINT DEFAULT 1
 );
