+++ src/main/resources/db/migration/V1__create-table-escolaridade.sql
@@ -1,9 +1,9 @@
 -- Criação da tabela de escolaridade
 CREATE TABLE escolaridade
 (
-    id_escolaridade INT AUTO_INCREMENT PRIMARY KEY,
+    id_escolaridade BIGINT AUTO_INCREMENT PRIMARY KEY,
     descricao       VARCHAR(255) NOT NULL,
-    is_active       TINYINT(1) DEFAULT 1
+    is_active       TINYINT DEFAULT 1
 );
