+++ src/main/resources/db/migration/V2__create-table-estado_civil.sql
@@ -1,8 +1,8 @@
 CREATE TABLE estado_civil
 (
-    id_estado_civil INT AUTO_INCREMENT PRIMARY KEY,
+    id_estado_civil BIGINT AUTO_INCREMENT PRIMARY KEY,
     descricao       VARCHAR(255) NOT NULL,
-    is_active       TINYINT(1) DEFAULT 1
+    is_active       TINYINT DEFAULT 1
 );
