+++ src/main/resources/db/migration/V4__create-table-permissao_usuario.sql
@@ -1,8 +1,8 @@
 -- Criação da tabela de permissão
 CREATE TABLE permissao_usuario
 (
-    id_permissao_usuario INT AUTO_INCREMENT PRIMARY KEY,
+    id_permissao_usuario BIGINT AUTO_INCREMENT PRIMARY KEY,
     nome                 VARCHAR(255) NOT NULL,
     descricao            VARCHAR(255),
-    is_active            TINYINT(1) DEFAULT 1
+    is_active            TINYINT DEFAULT 1
 );
