+++ src/main/resources/db/migration/V7__create-table-sexo.sql
@@ -1,9 +1,9 @@
 -- Criação da tabela de usuários
 CREATE TABLE sexo
 (
-    id_sexo   INT AUTO_INCREMENT PRIMARY KEY,
+    id_sexo   BIGINT AUTO_INCREMENT PRIMARY KEY,
     descricao VARCHAR(255) NOT NULL,
-    is_active TINYINT(1) DEFAULT 1
+    is_active TINYINT DEFAULT 1
 );
